Admin adjustments

+++ app/Http/Controllers/Admin/AdminController.php
@@ -2,20 +2,33 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Helpers\AccessLevel;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Admin\Admin;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Redirect;
 
 class AdminController extends Controller
 {
-    public function index(Admin $admin)
+    public function index()
     {
+        if(AccessLevel::isSimpleAdmin()){
+            $admin = AccessLevel::isSimpleAdmin();
+            return redirect(route('admin.admins.show', compact('admin')));
+        }
+
         $admins = Admin::all();
         return view('admin.admin.index', compact('admins'));
     }
 
     public function create()
     {
+        if(AccessLevel::isSimpleAdmin()){
+            return Redirect::back();
+        }
+
         return view('admin.admin.create');
     }
 
@@ -25,7 +38,7 @@ class AdminController extends Controller
 
         $admin->name = $request->admin['name'];
         $admin->email = $request->admin['email'];
-        $admin->password = $request->admin['password'];
+        $admin->password = Hash::make($request->admin['password']);
         $admin->save();
 
         return redirect(route('admin.admins.show', compact('admin')));
@@ -33,11 +46,18 @@ class AdminController extends Controller
 
     public function edit(Admin $admin)
     {
+        if(AccessLevel::isSimpleAdmin()){
+            $admin = Admin::where('id', AccessLevel::isSimpleAdmin())->first();
+        }
         return view('admin.admin.edit', compact('admin'));
     }
 
     public function destroy(Admin $admin)
     {
+        if(AccessLevel::isSimpleAdmin()){
+            return Redirect::back();
+        }
+
         $admin->delete();
         return redirect(route('admin.admins.index'));
     }
@@ -48,7 +68,7 @@ class AdminController extends Controller
 
         $admin->name = $request->admin['name'];
         $admin->email = $request->admin['email'];
-        $admin->password = $request->admin['password'];
+        $admin->admin_password = isset($request->admin['password']) ? Hash::make($request->admin['password']) : null;
 
         $admin->update();
 
@@ -57,6 +77,11 @@ class AdminController extends Controller
 
     public function show(Admin $admin)
     {
+        if(AccessLevel::isSimpleAdmin()){
+            $admin = Admin::where('id', AccessLevel::isSimpleAdmin())->first();
+            return view('admin.admin.show', compact('admin'));
+        }
+
         return view('admin.admin.show', compact('admin'));
     }
 
@@ -64,7 +89,8 @@ class AdminController extends Controller
     {
         $request->validate([
            'admin.name'       => 'required|min:4|max:50',
-           'admin.email'      => 'required',
+           'admin.email'      => $request->isMethod('post') ? 'required|email|unique:admins,email' : 'required|email|unique:admins,email,'.Auth::guard('admin')->user()->id,
+           'admin.password'   => $request->isMethod('post') ? 'required|min:8' : 'nullable',
        ]);
     }
 }
